Added baby gripping indication
Feature:
Whenever the pressure sensor is being applied a force at, the graph at the plot will signalize that by highliting in light gray the area.

How it works:
Basically there are two flags. One, the changeBackgroundColor, indicates that the graph should be highlighted. The other, renew, indicates that the a new AreaSeries should be added.

- Possible improovements: change the threshold for the pressure sensor or add more values.

CORRECTED BUGS:

- At line 405, it was corrected how it keeps track of which series to addPoint,  since it was occasionally causing overflow.

-