Added recording capabilities to pressure sensor thread. Testing new feature.
NEW FEATURE:
Now, everytime the pressure sensor detects a grip or release of the baby, the data will be recorded accordingly. That means, stop recording when the baby is released and centering and start recording when the baby is gripped.

-NEW FEATURE (NOT FINISHED):
Started to implement record everything feature. When the flag RECORD_EVERYTHING is true, instead of recording when the baby is gripped, the app will only start and stop recording data if the switch is manually  toggled on or off. Thus, the pressure sensor, instead of start recording, will signalize in the csv the start or end of pressure being applied. "+++" means the sensor started to be pressed and "---" the pressure was released.