Solving old known bugs
Solved a few crash occurencies, present since version 1 of ATT

-Battery check BUG:
When the back button in the linegraph page was pressed, the board variable was disposed but the battery check thread wasn't. That means that it would crash trying to check battery level without being possible.
Solution: dispose battery check thread.

-Initialize board Exception:
When going from the scanner page to the linegraph page a timeout exception can happen.
Solution: I added an exception check and small popup telling the user what happened. The try again button is just a dummy for now.