Design change in version 3 of the ATT
HOW WAS THE DESIGN UNTIL NOW:
The att version 3 until now was counting with a scroll view with 2 graphs and data showing about the sensors:
-One graph with euler angles Data
-Another graph with the axis angle and some quaternion raw data
This design has been changed

THE NEW DESIGN (unfinished):
In the new design the user will be able to alternate between the data he wants to be shown in the screen (euler or axis angle) with the use of the toggle buttons under the graph. For that to be done, the back button has reduced in size.

The new design isn't totally functional yet. The graph with euler data is hidden behind the axis angle graph for now and not accessible. The idea is that both graphs can alternate with the toggle buttons.

NOTICE:
This design change demanded lots of code change which for now may cause bugs from reminiscent code. With future test those bugs will be eliminated, but for now none was found.

LINES OF CODE CHANGED IN THE CS:
The changes in the xaml are self explanatory, but they cause some changes in the cs. Those were:

-Removal of the textblocks for the euler data. The euler data is now displayed in the same text blocks from the axis graph data when enabled.
- Now only plot method is executed depending on which graph is hidden
- The graph highlighting feature now should be functional to both graphs as opposed to how it was before when it was done only for the axis graph.